+++ packages/server/plugin-server/src/cli.ts
@@ -7,7 +7,7 @@ import type { AppTools } from '@modern-js/app-tools';
 
 const TS_CONFIG_FILENAME = 'tsconfig.json';
 
-export default (): CliPlugin<AppTools> => ({
+export const serverPlugin = (): CliPlugin<AppTools> => ({
   name: '@modern-js/plugin-server',
 
   setup: api => ({
@@ -60,3 +60,5 @@ export default (): CliPlugin<AppTools> => ({
     },
   }),
 });
+
+export default serverPlugin;