+++ packages/solutions/monorepo-tools/src/index.ts
@@ -1,6 +1,6 @@
 import type { CliPlugin } from '@modern-js/core';
-import changesetPlugin from '@modern-js/plugin-changeset';
-import lintPlugin from '@modern-js/plugin-lint';
+import { changesetPlugin } from '@modern-js/plugin-changeset';
+import { lintPlugin } from '@modern-js/plugin-lint';
 import { Import } from '@modern-js/utils';
 import { getLocaleLanguage } from '@modern-js/plugin-i18n/language-detector';
 import { i18n } from './locale';
@@ -15,7 +15,7 @@ const upgradeModel: typeof import('@modern-js/upgrade') = Import.lazy(
   require,
 );
 
-export default (): CliPlugin<MonorepoTools> => ({
+export const monorepoTools = (): CliPlugin<MonorepoTools> => ({
   name: '@modern-js/monorepo-tools',
   usePlugins: [changesetPlugin(), lintPlugin()],
   registerHook: hooks,
@@ -40,3 +40,5 @@ export default (): CliPlugin<MonorepoTools> => ({
   },
   post: ['@modern-js/plugin-changeset'],
 });
+
+export default monorepoTools;