+++ tests/integration/ssg/fixtures/nested-routes/modern.config.ts
@@ -1,5 +1,5 @@
-import appTools, { defineConfig } from '@modern-js/app-tools';
-import ssgPlugin from '@modern-js/plugin-ssg';
+import { appTools, defineConfig } from '@modern-js/app-tools';
+import { ssgPlugin } from '@modern-js/plugin-ssg';
 
 export default defineConfig({
   runtime: {