+++ tests/integration/swc/fixtures/transform-fail/modern.config.ts
@@ -1,4 +1,4 @@
-import appTools, { defineConfig } from '@modern-js/app-tools';
+import { appTools, defineConfig } from '@modern-js/app-tools';
 import pluginSwc from '@modern-js/plugin-swc';
 
 export default defineConfig({