+++ packages/generator/generator-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/generator-common
 
+## 3.1.24
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.24.0
+
 ## 3.1.23
 
 ### Patch Changes