+++ packages/generator/generator-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/generator-plugin
 
+## 3.1.24
+
+### Patch Changes
+
+- Updated dependencies [c882fbd]
+- Updated dependencies [4a82c3b]
+  - @modern-js/utils@2.24.0
+  - @modern-js/plugin-i18n@2.24.0
+  - @modern-js/generator-common@3.1.24
+  - @modern-js/generator-utils@3.1.24
+  - @modern-js/new-action@2.24.0
+
 ## 3.1.23
 
 ### Patch Changes