+++ packages/generator/generators/base-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/base-generator
 
+## 3.1.24
+
 ## 3.1.23
 
 ## 3.1.22