+++ packages/runtime/plugin-garfish/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/plugin-garfish
 
+## 2.24.0
+
+### Patch Changes
+
+- Updated dependencies [c882fbd]
+- Updated dependencies [56693a3]
+- Updated dependencies [4a82c3b]
+  - @modern-js/runtime@2.24.0
+  - @modern-js/utils@2.24.0
+
 ## 2.23.1
 
 ### Patch Changes