+++ packages/server/plugin-worker/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/plugin-worker
 
+## 2.24.0
+
+### Patch Changes
+
+- Updated dependencies [c882fbd]
+- Updated dependencies [36f5bdf]
+- Updated dependencies [4a82c3b]
+  - @modern-js/utils@2.24.0
+  - @modern-js/server-utils@2.24.0
+  - @modern-js/prod-server@2.24.0
+
 ## 2.23.1
 
 ### Patch Changes