+++ packages/server/prod-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/prod-server
 
+## 2.24.0
+
+### Patch Changes
+
+- Updated dependencies [c882fbd]
+- Updated dependencies [4a82c3b]
+  - @modern-js/utils@2.24.0
+  - @modern-js/server-core@2.24.0
+  - @modern-js/plugin@2.24.0
+
 ## 2.23.1
 
 ### Patch Changes